requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
boto3==1.34.79
botocore==1.34.79
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.1
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-ses==3.5.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
hashids==1.3.1
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
